fix(cli): Add entry point for CLI
It seems we forgot to add an entry point where we originally created the CLI.  This should fix it.

I also realized that the command doesn't give any indication of the simulation progress so I am adding a print statement so that people aren't left hanging for over a minute without any indication of how much longer the simulation will take.

We should ass an option to silence this reporting for those who don't want it.

CC: @saeranv