fix(zone): Include support for ClimateZones 1B and 5C
I assume that these climate zones were not originally included because there is no part of the contiguous US that contains them. However, they definitely exist on Earth and the ASHRAE 90.1 guidelines for them are effectively the same as other climate zones. So I am embedding this mapping with this commit.